Better Docs + Insertion Sort
I changed some of the wording basically everywhere, and tried to improve the documentation in areas I found it lacking.

I also tried my hand at making insertion sort using QTA and WOW, I did not expect it to be so difficult. This processor can probably do little niche projects, but even insertion sort is enough for you to seriously run the risk of running out of memory. Probably won't come back to this project now that I realize this, though it has me thinking about what a better version of the QPU would look like.